backend/requirements.txt
openai>=1.50.0
fastapi>=0.111
pydantic>=2.7
uvicorn>=0.30
httpx>=0.27
python-dotenv>=1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.0
chromadb>=0.4.0
pillow>=10.0.0
ddgs>=5.0.0
rembg>=2.0.50
sqlalchemy>=2.0.0
